src/profiles_project/requirements.txt
appdirs==1.4.3
codegen==1.0
Django==1.11
djangorestframework==3.6.2
packaging==16.8
patterns==0.3
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
